Extracted animations

+++ Project1/Source/IndexBuffer.h
@@ -3,7 +3,7 @@
 class IndexBuffer : public Bindable
 {
 public:
-	IndexBuffer(Graphics& gfx, const std::vector<unsigned short>& indices);
+	IndexBuffer(Graphics& gfx, const std::vector<int>& indices);
 	void Bind(Graphics& gfx) noexcept override;
 	UINT GetCount() const noexcept;
 protected:
